# generated/swagger.json
{
  "openapi": "3.0.0",
  "components": {
...
    "schemas": {
...
      "IApiValidationError": {
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": ["field", "message"],
        "type": "object",
        "additionalProperties": false
      },
      "Record_string.unknown_": {
        "properties": {},
        "additionalProperties": {},
        "type": "object",
        "description": "Construct a type with a set of properties K of type T"
      },
      "TErrorCause": {
        "$ref": "#/components/schemas/Record_string.unknown_"
      },
      "IHttpErrorDto": {
        "properties": {
          "name": {
            "type": "string"
          },
          "errorKey": {
            "type": "string"
          },
          "validationErrors": {
            "items": {
              "$ref": "#/components/schemas/IApiValidationError"
            },
            "type": "array"
          },
          "cause": {
            "$ref": "#/components/schemas/TErrorCause"
          }
        },
        "required": ["name", "errorKey"],
        "type": "object",
        "additionalProperties": false
      },
...
      "IPaginationResponseMeta": {
        "properties": {
          "page": {
            "type": "number",
            "format": "double"
          },
          "limit": {
            "type": "number",
            "format": "double"
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "totalPages": {
            "type": "number",
            "format": "double"
          }
        },
        "required": ["page", "limit", "total", "totalPages"],
        "type": "object",
        "additionalProperties": false
      },
...
      "UPaginationSortDirection": {
        "type": "string",
        "enum": ["asc", "desc"]
      },
...
      "UCounterpartyStatus": {
        "type": "string",
        "enum": ["active", "archived", "draft"]
      },
...
      "UCounterpartyType": {
        "type": "string",
        "enum": ["individual", "organization"]
      },
...
      "UCounterpartyRole": {
        "type": "string",
        "enum": ["employer", "vendor", "contractor"]
      },
      "IAddressDto": {
        "properties": {
          "line1": {
            "type": "string"
          },
          "line2": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          },
          "countryCode": {
            "type": "string"
          }
        },
        "required": ["line1", "city", "countryCode"],
        "type": "object",
        "additionalProperties": false
      },
      "ICounterpartyMetaDto": {
        "properties": {
          "employer": {
            "properties": {
              "address": {
                "$ref": "#/components/schemas/IAddressDto"
              },
              "displayName": {
                "type": "string",
                "nullable": true
              }
            },
            "required": ["address", "displayName"],
            "type": "object"
          },
          "vendor": {
            "properties": {
              "address": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/IAddressDto"
                  }
                ],
                "nullable": true
              }
            },
            "required": ["address"],
            "type": "object"
          },
          "contractor": {
            "properties": {
              "address": {
                "$ref": "#/components/schemas/IAddressDto"
              }
            },
            "required": ["address"],
            "type": "object"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
      "ICounterpartyDto": {
        "properties": {
          "version": {
            "type": "number",
            "format": "double"
          },
          "createdBy": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "accountingEntityId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/UCounterpartyStatus"
          },
          "type": {
            "$ref": "#/components/schemas/UCounterpartyType"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/UCounterpartyRole"
            },
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ICounterpartyMetaDto"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "version",
          "createdBy",
          "id",
          "accountingEntityId",
          "name",
          "status",
          "type",
          "roles",
          "meta",
          "createdAt",
          "updatedAt"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "IPaginatedResponse_ICounterpartyDto_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ICounterpartyDto"
            },
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/IPaginationResponseMeta"
          }
        },
        "required": ["data", "meta"],
        "type": "object",
        "additionalProperties": false
      },
      "UCounterpartySortBy": {
        "type": "string",
        "enum": ["createdAt", "name"]
      },
...
      "ICounterpartyCreateMetaReq": {
        "properties": {
          "employer": {
            "properties": {
              "address": {
                "$ref": "#/components/schemas/IAddressDto"
              },
              "displayName": {
                "type": "string",
                "nullable": true
              }
            },
            "required": ["address"],
            "type": "object"
          },
          "vendor": {
            "properties": {
              "address": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/IAddressDto"
                  }
                ],
                "nullable": true
              }
            },
            "type": "object"
          },
          "contractor": {
            "properties": {
              "address": {
                "$ref": "#/components/schemas/IAddressDto"
              }
            },
            "required": ["address"],
            "type": "object"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
      "ICounterpartyCreateReq": {
        "properties": {
          "name": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/UCounterpartyStatus"
          },
          "type": {
            "$ref": "#/components/schemas/UCounterpartyType"
          },
          "meta": {
            "$ref": "#/components/schemas/ICounterpartyCreateMetaReq"
          }
        },
        "required": ["name", "status", "type"],
        "type": "object",
        "additionalProperties": false
      },
      "ICounterpartyUpdateReq": {
        "properties": {
          "expectedVersion": {
            "type": "number",
            "format": "double",
            "description": "Version returned by the last read; stale writes fail with HTTP 409."
          },
          "name": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/UCounterpartyType"
          },
          "status": {
            "type": "string",
            "enum": ["active"],
            "nullable": false,
            "description": "Omit for ordinary edits; Active explicitly requests Draft activation."
          },
          "meta": {
            "$ref": "#/components/schemas/ICounterpartyCreateMetaReq",
            "description": "Complete role replacement when supplied; {} clears all roles."
          }
        },
        "required": ["expectedVersion"],
        "type": "object",
        "additionalProperties": false
      },
      "ICounterpartyArchiveReq": {
        "properties": {
          "expectedVersion": {
            "type": "number",
            "format": "double",
            "description": "Version returned by the last read; stale writes fail with HTTP 409."
          }
        },
        "required": ["expectedVersion"],
        "type": "object",
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "info": {
    "title": "drimsheet-core",
    "version": "0.1.0-alpha.1",
    "description": "This repo contains the server-side source code for Drimsheet's bookkeeping and tax app.",
    "contact": {
      "name": "Osahon Oboite"
    }
  },
  "paths": {
...
    "/counterparties": {
      "get": {
        "operationId": "getCounterparties",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IPaginatedResponse_ICounterpartyDto_"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          }
        },
        "description": "Get counterparties",
        "tags": ["Counterparty"],
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "orderBy",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sortDirection",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/UPaginationSortDirection"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "roles",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/UCounterpartyRole"
              }
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/UCounterpartyType"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/UCounterpartyStatus"
            }
          },
          {
            "in": "query",
            "name": "orderBy",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/UCounterpartySortBy"
            }
          }
        ]
      },
      "post": {
        "operationId": "createCounterparty",
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ICounterpartyDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          }
        },
        "description": "Create a new counterparty",
        "tags": ["Counterparty"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ICounterpartyCreateReq"
              }
            }
          }
        }
      }
    },
    "/counterparties/{id}": {
      "get": {
        "operationId": "getCounterparty",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ICounterpartyDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          }
        },
        "description": "Get a counterparty by id",
        "tags": ["Counterparty"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "patch": {
        "operationId": "updateCounterparty",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ICounterpartyDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          }
        },
        "description": "Update details and optionally activate a Draft. Supplied metadata replaces all roles.\nA type change after transaction use returns 409 with field=type,\nreason=transaction_usage and nextAction=create_counterparty.\nThis counterparty's type cannot be changed because it has been used in a\ntransaction. Create a new counterparty if a different type is required.",
        "tags": ["Counterparty"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ICounterpartyUpdateReq"
              }
            }
          }
        }
      }
    },
    "/counterparties/{id}/archive": {
      "post": {
        "operationId": "archiveCounterparty",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ICounterpartyDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          }
        },
        "description": "Archive a counterparty while retaining its details and transaction references.\nAn already archived counterparty returns unchanged without another history or event.",
        "tags": ["Counterparty"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ICounterpartyArchiveReq"
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}